[BUG] Allow for Parquet reading from files with differing schemas (#2514)
Fixes to allow for reading Parquet files and specifying **columns that
do not exist in the Parquet file**.

This is common when we try to "apply" a schema from some external source
(e.g. a data catalog). In that case, old Parquet files may not have
certain columns because the schema evolved over time. We want to make
sure that reads still succeed on these files, and we get back tables
with the appropriate number of rows (even if no columns were read!)

## Summary of Changes

**Fixes to Parquet reader**
1. Allows for reads of Parquet files with `column` names that may not
exist in the file. These columns will just be missing from the returned
`Table`. Note that this potentially means we get empty `Tables` with
valid number of rows.

**Fixes on `Table`**

1. Fixes `Table` to allow empty (num_rows=0) tables that still have
columns of data. This lets us do reads of files without producing any
data at all (e.g. reading only `col("x")` from a file that doesn't have
`col("x")`).
2. During reading of Parquet files, remove our `FieldNotFound` errors.
We no longer complain about a user-provided field not found, and will
just return `Table` structs without the requested columns instead
3. Fix `Table::new` to not default to num_rows=1. Instead, we pass in
num_rows explicitly and check against that.
4. Fix `Table::from_columns` to receive an explicit num_rows as well.
This cleans up a host of bugs where we might be naively creating tables
with the wrong length when a table has no columns.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>