[FEAT] Support Reading Iceberg Merge-on-Read Position Deletes (#2563)
When position deletes are present, I chose to disable row-group
splitting since it would complicate the logic to implement and it's
unlikely that people are storing a small number of very large individual
files in Iceberg. If there is a clean way to do this I wouldn't be
opposed to implementing it though.

This PR also does a slight refactor of the `DataFileSource` enum to
consolidate `AnonymousDataFile` and `CatalogDataFile` and remove
parameters that aren't relevant to database sources. In a future PR, I
would like to do some refactoring of our `ScanTask` struct in order to
ensure some relationships between the data sources and file formats. For
example a database format would not make sense to exist with a parquet
source, but we implicitly allow it right now. This has already caused
parts of our code to be messy and make unnecessary checks