[BUG] Fix `.str.length()` on Unicode strings (#2579)
Previously, the `.str.length()` method would count the number of bytes
in the UTF-8 string. This is inconsistent with Python's `len()` and
pandas' `str.len()` which count Unicode codepoints. For instance, on the
string "😉test", the number of bytes is 8, whereas the number of
codepoints is 5. This PR makes Daft consistent with that behavior.

There doesn't seem to be a way now to reproduce the original behavior;
maybe we should add a `.byte_length()` method for that.