[FEAT] Enable buffered iteration on plans (#2566)
Helps close part of #2561 

This PR enables buffering of result partition tasks, preventing "runaway
execution" of executions when run concurrently.

The problem previously was that if we ran two executions in parallel
(`e1` and `e2`) on a machine with 8 CPUs:

1. `e1` could potentially run `8` tasks and keep them buffered (not
releasing the resource request)
2. When `e2` attempts to run the next task, it notices that the task
cannot be admitted on the system (due to memory constraints)
* `e2` thinks that it is deadlocking because there is a strong
assumption in the pyrunner today that if a task cannot be admitted, it
merely has to wait for some other tasks in the same execution to finish
up.
* However, `e2` doesn't have any tasks currently pending (because it is
starved). The pending tasks are all buffered in `e1`. Thus it thinks
that it is deadlocking.

## Solution

* This PR sets the default buffering behavior to `1` instead of allowing
each execution to run as many tasks as it wants
* We introduce logic in the physical plan to have an upper limit on the
size of the materialization buffer. If that buffer gets too large, it
will start yielding `None` to indicate that the plan is unable to
proceed.

Note that there is still potentially a problem here, e.g. running >
NUM_CPU number of executions concurrently. That can be solved in a
follow-up PR for refactoring the way we do resource accounting.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>
Co-authored-by: Desmond Cheong <desmondcheongzx@gmail.com>