[BUG] Fix parquet reads when a top-level column's final row spans more than one data page (#2586)
This PR addresses issue #2537 by fixing parquet reads in the case where
the final row of a top-level column spans more than one data page.

Previously, it was implicitly assumed that all nested values for a
top-level row was stored in the same data page. So once we had a nested
value for each top-level row, we assumed that it meant we had read the
final data page for a column.

However, this assumption isn't guaranteed in parquet. In the event that
the final row of a top-level column spans multiple data pages, we stop
reading data pages for the column at the first page where we first
encountered the row. This could either lead to data loss (we stopped
reading nested data early) or cause Daft to error out (in the case where
the nested fields are in a struct for example, one field having a
different number of values from other fields causes Daft to throw an
error).

To fix this, in addition to checking that we've seen every row before we
stop reading data pages for a column, we also check the number of values
we've read against the number of values we expect from the column's
metadata.

### Example

Consider the following set up:
- We have a dataframe with the schema `<nested list<struct<field1
string, field2 string>>>`.
- The strings in `nested.field1` are large and unique enough that
dictionary encoding falls back to plain encoding.
- Each data page stores at most 1024 of `nested.field1`'s strings.
- We have two top-level rows in our dataframe, with the first row
containing a list of length 1, and the second row containing a list of
length 3072.

This gives us the following column layout within the parquet file:
```
Column: nested.list.item.field1
--------------------------------------------------------------------------------
  page   type  enc  count   avg size   size       rows     nulls   min / max
  0-D    dict  S _  1024    2.931 kB   2.931 MB  
  0-1    data  S R  1024    1.27 B     1.270 kB            0       "a" / "zzeirptdzfmtwpousxsenxqaz..."
  0-2    data  S _  1024    2.934 kB   2.934 MB            0       "aaddwlmkjpjkueymxhffkwvqk..." / "zzhwmmqrhngwpqbusvfdbrzqw..."
  0-3    data  S _  1024    2.934 kB   2.934 MB            0       "aakrkhbhhknfwffdlisrbymvf..." / "zwuuxcjwkdxkbkjrehcriurzt..."
  0-4    data  S _  1       2.945 kB   2.945 kB            0       "brweggkctrffxwkkfcpkegfzz..." / "brweggkctrffxwkkfcpkegfzz..."


Column: nested.list.item.field2
--------------------------------------------------------------------------------
  page   type  enc  count   avg size   size       rows     nulls   min / max
  0-D    dict  S _  2       5.00 B     10 B      
  0-1    data  S R  3073    0.01 B     22 B                0       "b" / "r"

```

Since the last row of `nested.field` contains 3072 values, it overflows
page `0-1` and goes into pages `0-2`, `0-3`, and `0-4`.

If we stop reading data pages once we've seen a value from every
top-level row in a column, then we would only read column `nested.field`
up to page `0-1`, and miss the next three pages of data. Instead, we can
use the column's metadata so that we know to expect 3072 values and
continue processing data pages until we've seen all these values.