[FEAT] Make the `end` argument for `.list.slice()` optional (#2593)
Currently, to use the `.list.slice()` expression, `start` and `end`
arguments must be specified to determine which indices of the list to
slice.

Since it's convenient to slice from a starting index to the end of the
list without needing to compute how long the list is, this PR makes
`end` optional. When `end` is unspecified, `.list.slice()` slices the
list from `start` until its final element.