[FEAT] Add ActorPoolProject logical and physical plans (#2601)
Adds new LogicalPlan and PhysicalPlan nodes for ActorPoolProject

Note that this is not yet reachable via user-facing code, since nothing
from our API should be able to create logical `ActorPoolProject` plans
yet (this will be left to a follow-on PR and we can feature-flag it)

---

**ActorPoolProject** works just like a normal Project (it has a list of
expressions that it is responsible for evaluating). The main differences
are:

1. The scheduler is responsible for running these projections on top of
an actor pool, instead of naively running it like a normal projection
2. The expressions are expected to contain `PyPartialUDFs`, which the
execution backend is responsible for initializing once-per-actor.
3. (TODO in a follow-up PR) During evaluation of the expressions, we
will be looking for the initialized Python classes from some global
state. The projections in an ActorPoolProject cannot be evaluated
without that global state having been initialized.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>
Co-authored-by: Desmond Cheong <desmondcheongzx@gmail.com>