[BUG] Fix projection pushdowns not working with limits (#2635)
Fixes #2616

From Slack:
> Essentially the project and the limit are being pushed down in the
same rule batch, at the same time. However, this means that if a project
and a limit are in the same tree, the project gets pushed down, but then
the limit gets immediately pushed down beneath it, resulting in no
change, and the project never gets to merge with the scan.

This PR moves limit pushdowns into a different rule batch than
projection pushdowns, meaning they no longer conflict with each other.