[BUG] Fix join op names and join key definition (#2631)
This simplifies our logic for joins by making the right side renaming an
explicit project before the join during logical planning. This allows
code further in the pipeline to simply assume that the column names are
preserved during the join and that the only matching column names
between the left and right sides are the join keys. This PR also moves
the revised `infer_join_schema` function into `daft-dsl` so that it can
be used both in planning and execution.

Moreover, this PR changes join behavior slightly, by only merging join
keys from both sides if they are exact column expressions. For example,
before `df1.join(df2, left_on="a", right_on=col("a") + col("b"),
how="outer")` would have created a single column "a" with both the left
and right values, but now does not, and instead splits it into "a" and
"right.a".

This does not fix the filter pushdown bug with joins, which will be
tackled in a later PR.

Also resolves #1294