[BUG] Fix nested parquet reads for .show() and .limit() (#2643)
#2586 did not consider the case when `.show()` or `.limit()` was used.
In these cases, the result is typically truncated to a few rows of data.
With the fix in #2586, when calling `.show()` or `.limit()` on a parquet
file with a `list` column, our parquet reader would read every data page
until all values have been read despite only needing a few data pages.
This gives result Series that have mismatching lengths, e.g. one might
see an error: `daft.exceptions.DaftCoreException: DaftError::ValueError
While building a Table with Table::new_with_size, we found that the
Series lengths did not match. Series named: line_ids had length: 23054
vs the specified Table length: 8`.

This PR fixes the issue by zeroing out the number of values left to read
in a nested column once we know that we have seen all values in the
requested rows. To do this, we either have to read all values (in the
case where the user wants to read all rows), or we have to read a
parquet repetition level of `0` (which indicates the start of a new
record) after we have encountered the desired number of rows.