[BUG] Fix filter pushdown into non-inner joins (#2659)
There are some additional optimizations I would like to make but I
decided to keep this PR minimal. Aside from splitting a filter into its
conjunctions to allow for partial predicate pushdowns, this PR does the
same optimizations as the original when it is correct to do so.

Also a flyby change on some `docker-compose.yaml` files to remove some
image specifiers that were causing some failed tests and were probably
meant to actually describe the name of the service?