[FEAT] Allow user provided schema and schema inference length for read_sql (#2676)
Enables users to provide a `schema` to read_sql, as well as an
`infer_schema` flag to turn on/off inference, and `infer_schema_length`
to customize the length of the scan for inference.

Note that `infer_schema_length` only applies when SQLAlchemy is used to
read, otherwise ConnectorX is able to provide the schema with a
`limit=0` as it is an arrow backed driver. This is a new optimization.

This PR also has a few refactors, namely moving the
`construct_sql_query` and the `read_schema` logic to the SQLConnection
object instead of the ScanOperator. This is so that schema inference
also has context on the driver that will be used.

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>