[PERF] Fix excessive parquet metadata reading (#2694)
When reading files from AWS, the logical to physical plan translator
previously fetched the metadata for the files sequentially, which could
be very slow if there are many files. (This bug was introduced in
#2358.)

This PR makes it so that we only cache the metadata upon splitting, and
when we do so we only cache the row groups that are actually relevant to
each scan task. This avoids serializing the entire metadata for each Ray
runner, which should improve performance.

Benchmark results:
<img width="516" alt="image"
src="https://github.com/user-attachments/assets/ba013482-89be-413f-89da-8f0e8fcf4cd7">