[CHORE] Add check for stateful UDF outside of project (#2771)
Add a parameter to `resolve_expr` and its related functions to allow for
stateful UDFs in the expression. This is set to false everywhere except
for projects.

Additionally, `df['column name']` used to call `resolve_exprs` to check
for validity and returned its output, which would be problematic if
there were wildcards in the expression. Now, I've created a function
that only does the validity check, and `DataFrame.__getitem__` would
just return `col(name)`, which will be actually resolved later in the
builder.