[CHORE] Use treenode for tree traversal in logical optimizer rules (#2797)
Follow-up from #2791 after offline conversation.

This PR still uses the existing logical plan optimizer, but now uses
`common_treenode::Transformed` instead of a custom implementation. In
addition, it removes the apply order logic from the optimizer and
replaces it with `common_treenode::TreeNode` transforms done in the
optimizer rules themselves.

This PR should not cause any functional changes to any of the optimizer
rules or the way they are applied, except for the fact that rules in a
batch are now each applied to the whole tree before the next rule,
instead of each being applied to a single plan node if they have the
same apply order. Future PRs will separate the rules out and make better
use of treenode.