[CHORE] Remove user-facing arguments for casting to Ray's tensor type (#2802)
## Summary

Cleanup PR.

1. Removes `cast_tensors_to_ray_tensor_dtype` as a user-facing argument
in our export methods (e.g. `to_arrow`, `to_pandas` etc) -- this is
really only intended to be used when a user is converting a Daft
dataframe to a Ray dataset anyways and there isn't a need to expose this
functionality to a user
2. Instead, the logic for casting `daft.DataType.tensor` data to a Ray
Data tensor type is done inside of the conversion code for Ray Data
(`_make_ray_block_from_micropartition`). This lets us contain the
ickiness of that code without having it touch all of our `to_arrow`
logic
3. Also removes `_trim_pyarrow_large_arrays` which was a legacy codepath
that doesn't get hit anymore

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>