[FEAT] [SQL] Add global agg support for SQL (#2799)
This adds support to SQL to run global aggregation when there are only
aggregations and no group by, which is inline with many other SQL
implementations.

It is similar in nature to
https://github.com/Eventual-Inc/Daft/issues/1979, but specific to SQL -
depending on how that issue is resolved the SQL planner code here could
be redundant.

Example:
```
d = daft.from_pydict({"i": [1,2,3,4,5]})
daft.sql('''SELECT sum(i) FROM d''', SQLCatalog({'d': d})).collect()
╭───────╮                                                                                                    
│ i     │
│ ---   │
│ Int64 │
╞═══════╡
│ 15    │
╰───────╯
```

*note*: I know the SQL support is new, so sorry if this is jumping the
gun! Is it better to first create discussions or issues for this kind of
change? I've been looking forward to SQL in this project and from
playing around with using sqlglot on the python side to build a logical
plan, I noticed a few things - now that this is here I thought I might
contribute 😄