[CHORE] Ensure compatibility with deltalake version v0.19 (#2827)
Deltalake v0.19 changes their `_convert_pa_schema_to_delta` function to
take in a `schema_conversion_mode` instead of `large_dtypes`. Pyarrow
also needed to be upgraded to v16.0.0 as well to be compatible with the
new version of deltalake

The difference between this PR and #2754 is that it still maintains
compatibility with older deltalake versions.

The reason why this PR also includes a change to arrow2 is because
starting in version 0.19, deltalake uses arrow-rs by default instead of
pyarrow to write files when calling `deltalake.write_deltalake`. We do
not actually use this functionality but our tests do, and arrow-rs
writes map arrays in a way that does not conform to the parquet spec. I
figured it would be good to just add that compatibility in there just in
case some user is using `arrow-rs` to write their parquet files.

However, there are also other issues with deltalake's rust writer,
including improper encoding of partitioned binary columns, so we will
use their pyarrow writer for testing.