[FEAT] Delta lake allow unsafe rename for local writes (#2824)
I don't know a good way to test this since we believe this issue only
arises when you mount a fabric table onto the local filesystem. However
it is a pretty change that should not affect any existing behavior. I
will verify that it works with our users once it is released.