[FEAT] Iceberg partitioned writes (#2842)
This also changes the behavior of some of the partitioning functions in
ways that I would consider as bug fixes. However @samster25 maybe you
should take a look at them to make sure their new behavior is correct.
The changes:
- truncate function now renames columns to `{}_trunc` instead of
`{}_truncate` to match Spark behavior
- day partitioning now returns an Int32Array instead of a DateArray. I
believe the past behavior was there to match pyiceberg, but talking to
the pyiceberg team, this actually seems like a bug. I plan on making a
PR to pyiceberg to fix this, but I have also moved and fixed the buggy
logic to our codebase so that it works with past versions of pyiceberg
as well.