[FEAT] Add Sparse Tensor logical type (#2722)
Closes #2494

It's still WIP, wanted to get some feedback on the draft and see if I am
in the right path.

Still having issues with:

- When casting From COOSparseTensorArray to TensorArray keeping the
values dynamically type. What I currently do is assuming the type in
order to be able to iterate over the values and in insert the non zero
values in the relevant indices.