[BUG] Pass parquet2 io errors correctly into arrow2 (#3012)
In
https://github.com/Eventual-Inc/Daft/commit/7fe3dbceb3ca3ee0c555753f0a55bd7e0c57bba6,
two changes were made that caused parquet2 io errors to be mishandled.

Firstly, a `IoError` for parquet2 was introduced. However in the
implementation of `parquet2::error::Error` for `arrow2::error:Error`,
any parquet2 error that is not `FeatureNotActive` or `Transport` is
transformed into an `ExternalFormat` error.

Secondly, arrow2 IO errors are intended to be marked as
`ByteStreamError`s, primarily so that they are handled as transient
errors and can be retried appropriately. However this special case was
removed.

This PR makes now classifies parquet2 io errors as arrow2 io errors, and
arrow2 io errors are now marked as `ByteStreeamError`s.

---------

Co-authored-by: Raunak Bhagat <rabhagat31@gmail.com>