[BUG] Fix intersection checking when unioning schemas (#3039)
In the definition of `Schema::union`, the error message suggests that we
intended to throw errors when performing a union on two schemas with
overlapping keys. However, the original implementation took the set
difference of keys between one side of the union and itself, which would
never throw an error.

This bug was not noticed because the python tests went through the
python code path which would check for the intersection correctly. But
if one uses the Rust API directly, then this property is not upheld.

We fix this bug by instead checking that the two sides of the union have
distinct keys.