[FEAT]: Interval dtype (#3018)
### Description
implement the arrow `interval` datatype in daft. 

#### Why? 
Our current `duration` dtype is useful for various operations, but it
only designed to handle **fixed** durations. Users will often want to
perform operations using relative times such as years, months, and
weeks, but this is currently only possible by incorporating additional
libraries to handle the relative to fixed conversions for us.

The addition of `Interval` means that we can **natively** operate on
relative times. This opens up a lot of possibilities for SQL, as we can
now use the SQL `INTERVAL` operator to operate on datetimes.



### Still TODO

- implement ops for Interval
  - [ ] add
  - [ ] sub
- [ ] python tests 

### Note for reviewer:
arrow's interval spec is kinda weird, and is actually 3 different
implementations depending on the interval

- years and months are represented as a logical type over i32
- days/millis is a physical type over `(i32, i32)`
- month/day/nano is a physical type over `(i32, i32, i64)`

Since most of the kernels are built around the `month/day/nano` variant,
I created the daft `Interval` type around that one. This introduces a
slight amount of overhead when using only year/month _(extra i32 & i64)_
or days/millis _(extra i64)_

But considering the most common pattern for interval is to do date math
on `Timestamp` columns, this should not be a huge concern.

---------

Co-authored-by: Colin Ho <chiuhong@usc.edu>