[CHORE] Enable lancedb reads for native executor (#2925)
Enables Lance reads for swordfish. 

The existing lance read mechanism reads from python iterators. This PR
abstracts out that logic to a separate function.

Both the python and streaming executors will call this function, the
python executor will do a `.collect` on the iterator, while the native
runner will wrap the iterator in a stream.

---------

Co-authored-by: Colin Ho <colinho@Colins-MBP.localdomain>
Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>