[BUG] Fix actor pool initialization in ray client mode (#3028)
This PR moves the `actor_pool_context` method from the ray runner to the
scheduler, and routes the relevant `ActorPoolManager` implementation to
`actor_pool_project`. That way, it does not accidentally get the wrong
actor pool context when the scheduler is running on a ray actor, which
we do when in ray client mode.

With this change, in addition to separating the `actor_pool_context`
method out of `Runner` into `ActorPoolManager`, I also move some other
things around to clean things up, especially so that `ray_runner.py` no
longer depends on the `pyrunner.py`.

Note:
- One of the tests was cut down because it needed 3 CPUs to run, which
combined with the 1 CPU for the scheduler actor, meant no more tasks
could even be scheduled. I plan on adding some more informative
errors/warnings when this happens in a future PR